✅ validate input (to some extent)
It is tempting to offload input JSON validation to the framework. Though
this approach has its drawbacks as well: it would be impossible to
recover from a malformed input data. The task requirements have no
mention on the sources of input data and their validity, so I think of
two scenarios:

* the best case if we control the source of input data (e.g. it is
  generated by our Web application frontend). Then the application could
  implement the most strict technique and fail on every minor malformed
  data;
* the worst case is when this data is supplied from a third-party, e.g.
  from a client. In this case we should [probably] do our best to recover
  from misalignments and report warnings to a user.

Also, framework-powered JSON validation turned out to be impossible to
perform fully, because of Hybernate validator problems when working with
Scala collections:

https://github.com/datasift/dropwizard-scala#limitations

Angry note. This is where Scala suck. It declares interoperability with
Java, but it is flawed every time. As you dig deeper, you inevitably
face this kind of issues.

Devil advocate here. Compare to Node.js, where you go to
http://yarnpkg.com/ or https://www.npmjs.com/ and easily find "yup"
or "joi" and both will cover all the JSON validation cases regardless
JavaScript flavor, being it TypeScript or Babel or Flow. Any library
will work seamlessly either with GraphQL or in a vanilla Express
application.