✨ ✅ normalize and transform an input for resource to serve it easily
Some notes on the motivation of rational choices in the algorithm
implementation. The balance is between:

* complexity. I am assuming the amount of input data is relatively
  small, so there is no need to optimize space/time-wise and better focus
  on simpler code and the next points;
* maintainability. The code should be easily understandable by others so
  that it can evolve;
* extensibility. I left enough comments and possibilities to implement a
  better error recovery later.

On the technology choice. I used validation in Scala since 2009 (note
who asked the question):

https://stackoverflow.com/questions/1193333/using-either-to-process-failures-in-scala-code

Regarding Scalaz and Cats, see the answers above. In my opinion, both of
them are a lot of technology and I've decided to keep it simpler here.

On the error recovery I have an experience documented as well:
https://github.com/alaz/slides-err-recovery